Coordinates Getter
A simple console program that allows the user to enter the desired
coordinates, which would be saved into a GeoJSON file and used to search
with Planet's API. For now, the code does not feature any checks for
properly input coordinates, except for the one that checks whether the
longtitude and latitude are bigget than zero. Should the user enter
invalid coordinates, he would get an unusable file.